feat: Adds memory limit support - Adds a new resource limiter which manages page allocation, indicating when limits have exceed and ensuring allocation cannot exceed limits - When a sandbox fails, the resource limited is checked for failed attmpts, indicating OOM errors. - Adds tests to verify behaviors.

+++ src/error.rs
@@ -15,7 +15,7 @@ pub enum SandboxError {
     OutOfFuel,
 
     #[error("guest exceeded its {limit_bytes}-byte memory limit")]
-    MemoryExceeded { limit_bytes: usize },
+    MemoryExceeded { limit_bytes: u64 },
 
     #[error("guest exceeded its execution timeout")]
     Timeout,
